Assignment2 - RestAPI with GAE and Cloud/requirements.txt
Flask==2.0.1
gunicorn==20.0.4
google-cloud-datastore==2.1.3
Werkzeug==2.2.2